fix: resolve size_t to int conversion warnings in test files
Fix remaining type conversion warnings in test files to achieve
completely clean compilation for both main code and tests.

**Changes:**

1. **WechatCallbackIntegrationTest.cc** - Fix 3 conversion warnings
   - Line 103: nonce.size() → static_cast<int>(nonce.size())
   - Line 124: aad.size() → static_cast<int>(aad.size())
   - Line 137: plaintext.size() → static_cast<int>(plaintext.size())

2. **WechatPayClientTest.cc** - Fix 3 conversion warnings
   - Line 35: nonce.size() → static_cast<int>(nonce.size())
   - Line 56: aad.size() → static_cast<int>(aad.size())
   - Line 69: plaintext.size() → static_cast<int>(plaintext.size())

**Root Cause:**
OpenSSL EVP functions expect int parameters but receive size_t
from std::string::size(), causing potential data loss warnings.

**Verification:**
- ✅ Zero warnings in main code compilation
- ✅ Zero warnings in test code compilation
- ✅ PayServer.exe builds successfully (6.6MB)
- ✅ PayBackendTests.exe builds successfully (10.6MB)

**Impact:**
- Complete clean compilation across entire project
- Type-safe cryptographic operations
- Professional code quality with zero compiler warnings

This completes the warning resolution effort started in the previous
commit, achieving a fully clean build process.

+++ PayBackend/test/WechatCallbackIntegrationTest.cc
@@ -100,7 +100,7 @@ std::string encryptAesGcm(const std::string &plaintext,
         return {};
     }
 
-    if (EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_GCM_SET_IVLEN, nonce.size(),
+    if (EVP_CIPHER_CTX_ctrl(ctx, EVP_CTRL_GCM_SET_IVLEN, static_cast<int>(nonce.size()),
                             nullptr) != 1)
     {
         EVP_CIPHER_CTX_free(ctx);
@@ -121,7 +121,7 @@ std::string encryptAesGcm(const std::string &plaintext,
     {
         if (EVP_EncryptUpdate(ctx, nullptr, &outLen,
                               reinterpret_cast<const unsigned char *>(aad.data()),
-                              aad.size()) != 1)
+                              static_cast<int>(aad.size())) != 1)
         {
             EVP_CIPHER_CTX_free(ctx);
             return {};
@@ -134,7 +134,7 @@ std::string encryptAesGcm(const std::string &plaintext,
             reinterpret_cast<unsigned char *>(&ciphertext[0]),
             &outLen,
             reinterpret_cast<const unsigned char *>(plaintext.data()),
-            plaintext.size()) != 1)
+            static_cast<int>(plaintext.size())) != 1)
     {
         EVP_CIPHER_CTX_free(ctx);
         return {};
